Generalize modular addition scripts
Improving modular addition scripts and adding a train config.

These scripts generalize the modulus.

According to the grokking paper, we should actually aim for a
(relatively) larger prime number for replication, such as 101.

Also to observe the synthesis costs, we need to increase printouts for
test and validation loss -- requiring customization of the training
config.