Refactor train.py to utilize argparse
This will make train.py more suitable for utilizing for exploration and
large queues of experiments.

This change also merges the following files into train.py:
- `configurator.py`
- `./config/train_shakespeare_char.py` and other config files, replacing
  these static configurations with more dynamic configuration ranges.

We safely remove the `configurator.py` file, but keep the static config
files for now, as we explore ideas for implementing the dynamic replacement.

Next steps are to simplify the means for us to run -- and also log and
act upon results results from -- larger experiments.