Add polymax
Add a parameterized estimation of softmax which should be easier to
implement in hardware.

This has three sections:

1. flat section (y=0)
2. linear section  (y=mx+b)
3. polynomial section (y=x**p + b), where 'p' is the polynomial power (e.g. 2)

To prevent overflow, we divide as done in the 'constantmax' variation by
a number, in this case by default 1000.0.

Preliminary experiments on smaller networks shows this converges for
power 2 and divisor of 1000.0, without having to implement an equivalent
for x_max.