If value is empty, do not add label or value
If for some types of data one has missing fields, then we do not want to
just have a hanging label. We want the label only to be added when there
is data afterwards.

The network will then learn that whenever there is a label, a valid
value will follow.

This allows modularity and later composability.

- Modularity - we can omit entire columns of the csv after initial
  training and the network should be able to adapt
- Composability - we can add new columns later and (in so far as the
  labels are allowed by the vocab space) we should be able to expand the
  network to train on the new values.