Refactor to trim train.py
Let's refactor to do visualizations from npy files, separate logic from
the train.py script (which will focus on training).

As part of this refactoring, adding a major checkpoint, which will be an
interval (if not None) at which a checkpoint will be saved regardless of
the validation loss.

This will be useful for monitoring how the weights and learned values
progress over time.