Add hypersphere norm variation
This is a slight variation which removes gain and thus
performs mapping of the token vector to the surface of a hypersphere.

As to the radius of the hypersphere, optional parameters are added for
setting the hypersphere radius directly (defaulting to square root of the
embedding dim), also optionally letting it be a learned after initializing.