Working version of multicontext embeddings
Appears to work with top k = 1, though the lower entropy embeddings
will be much more certain.

This suggests that the vector holds more information.

This solves congestion of the lm head
This solves decoding via the differentiation of buffalo (n) buffalo (v), and reduces repetition
This solves labelling of larger entities
This solves simultaneous data -- like tones for pinyin
This increases the speed of learning (higher loss function and more angles)
This improves the amount of learning from a fixed dataset.