Update CustomCharByteFallback to begin with bytes
Before we were ending with bytes, which makes it more difficult to
append more non byte tokens later, as well as making the boundary of
bytes to non-byte tokens variable.

This makes the code cleaner, and more extensible in case we wanted to
merge in more vocab later simplifying tracking of the byte-fallback
boundaries in the vocab.