Fix variable name alignment and if conditions
Aligns the sequential and parallel setup scripts to ensure that we check
the correct granularity of setting before we do the norm.

This also changes out_ln_attn -> peri_ln_attn, which will allow us to
keep the same naming structure.

Note that peri ln is only peri ln though when pre-norm is also
activated, creating the perimeter normalization sandwich around the
module.